% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{format_cols}
\alias{format_cols}
\title{Format columns for printing}
\usage{
format_cols(.data, .cols = NULL, ...)
}
\arguments{
\item{.data}{A \code{data.frame} or \code{tibble}.}

\item{.cols}{<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}> columns to format.
If \code{NULL} (the default), will format all \code{double} columns.}

\item{...}{
  Arguments passed on to \code{\link[=format]{format}}
  \describe{
    \item{\code{}}{}
  }}
}
\value{
An object of the same type as \code{.data} with any formatted columns
converted to \code{character} mode.
}
\description{
Wraps around \code{\link[base:format]{base::format()}} and \code{\link[dplyr:mutate]{dplyr::mutate()}} to format columns of a
\code{tibble} or \code{data.frame}. Note that any formatted columns will be converted
to \code{character} mode, so use this function only for printing.
}
\examples{
## Convert to tibble for better printing
mtcars_tbl <- tibble::as_tibble(mtcars)

## Formatting all columns
format_cols(mtcars_tbl, digits = 0)

## Specific columns
format_cols(mtcars_tbl, c("wt", "qsec"), digits = 0)

## Other formatting options
format_cols(mtcars_tbl, scientific = TRUE)

format_cols(mtcars_tbl, trim = TRUE)

}
\seealso{
\link[base:format]{format}
}
