% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{format_percentage}
\alias{format_percentage}
\title{Format numbers as percentages}
\usage{
format_percentage(x, ...)
}
\arguments{
\item{x}{a numeric vector to format.}

\item{...}{other arguments passed on to \code{\link[scales:label_percent]{scales::label_percent()}}}
}
\value{
A character vector of \code{length(x)}.
}
\description{
This is a simple wrapper around \code{\link[scales:label_percent]{scales::label_percent()}} to convert numeric
input to percentages and return them as characters.
}
\examples{
x <- seq(0, 1, length.out = 5)
format_percentage(x)

}
