% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{save_plots}
\alias{save_plots}
\title{Save multiple plots to a single PDF}
\usage{
save_plots(plot_list, filename)
}
\arguments{
\item{plot_list}{List of plot objects. Can be any type as long as
it has a \code{print()} output.}

\item{filename}{Character string. Can also be specified as path.}
}
\description{
Takes a list of plots and exports them to a single PDF with one plot per page.
}
\examples{
\dontrun{
  library(ggplot2)

  # Generate list of plots
  mtcars_split <- split(mtcars, mtcars$cyl)
  plot_list <- lapply(mtcars_split, function(d) {
    ggplot(d, aes(mpg, wt)) +
      geom_point()
  })

  # Save plots
  save_plots(plot_list, "plots.pdf")
}

}
