% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{gg_pval_hist}
\alias{gg_pval_hist}
\title{Create p-value histogram}
\usage{
gg_pval_hist(p_values)
}
\arguments{
\item{p_values}{Vector of p-values.}
}
\value{
A ggplot object.
}
\description{
Use \code{\link[ggplot2:geom_histogram]{ggplot2::geom_histogram()}} to create a histogram of p-values but with
better binwidths and axis limits.
}
\details{
By default, \code{\link[ggplot2:geom_histogram]{ggplot2::geom_histogram()}} centers the first and last bins of
the histogram on the x-axis lower and upper limits, respectively.
In case of p-values, which lie in the interval [0, 1], it makes more
sense to set the boundaries of the first and last bins at 0 and 1.
The binwidths are also set at 0.05, commonly used as a cut-off for significance,
so that the height of the first bar represents the number of p-values lying
between 0 and 0.05, i.e. the number of cases that would be deemed significant.
}
\examples{
pvals <- runif(n = 1000, min = 0, max = 1)
gg_pval_hist(pvals)

}
